Step 3
It's starting too look good. There are still a lot of unwraps and a few
not secured code flow paths that need to be fixed. The refresh tokens
need to be implemented for OAuth2 to be kinda working and `/userinfo`
should contain more data. Finally, I need to implement scopes as well
as user e-mails.